+++ src/components/ids-data-grid/ids-data-grid.ts
@@ -379,10 +379,10 @@ export default class IdsDataGrid extends Base {
 
           const headerHeight = this.header?.getBoundingClientRect?.().height ?? 0;
           this.container.scrollTop = scrollTopPixels - headerHeight;
-          this.scrollRowIntoView(rowStart);
+          this.#scrollRowIntoView(rowStart);
           requestAnimationTimeout(() => {
             this.#attachVirtualScrollEvent();
-            this.scrollRowIntoView(this.rowStart);
+            this.#scrollRowIntoView(this.rowStart);
           }, 150);
           handleReady();
         }
@@ -1447,7 +1447,7 @@ export default class IdsDataGrid extends Base {
       if (rowIndex === debounceRowIndex) return;
       debounceRowIndex = rowIndex;
 
-      this.scrollRowIntoView(rowIndex, false);
+      this.#scrollRowIntoView(rowIndex, false);
     }, { capture: true, passive: true }); // @see https://javascript.info/bubbling-and-capturing#capturing
   }
 
@@ -1495,39 +1495,24 @@ export default class IdsDataGrid extends Base {
    * @param {number} rowIndex row index
    */
   #scrollTo(rowIndex: number): void {
-    this.rowByIndex(rowIndex)?.scrollIntoView?.();
     const headerHeight = this.header.clientHeight;
-    const scrollHeight = this.container!.scrollHeight;
-    const containerHeight = this.container!.clientHeight;
-    const scrollTop = this.container!.scrollTop;
-    const isScrollBottom = (scrollTop + containerHeight) >= scrollHeight;
-
-    // offset for sticky header height
-    if (!isScrollBottom) this.container!.scrollTop -= headerHeight;
+    this.container!.scrollTop = (this.rowByIndex(rowIndex)?.offsetTop || 0) - headerHeight;
   }
 
   /**
-   * We always want to set doScroll=true when scrollRowIntoView() is called manually in code...
-   * ...so when the "public" uses it they would simply do scrollRowIntoView(x).
-   *
-   * However, this method is also used in the "onscroll" event-handler...
-   * ...within that "onscroll" event-handler, we want doScroll=false,
-   * ...and let the browser handle moving/panning the window without interference.
-   * @param {number} rowIndex - which row to scroll into view.
-   * @param {boolean} doScroll - set to "true" to have the browser perform the scroll action
-   * @see IdsDataGrid.#attachVirtualScrollEvent()
-   * @see https://medium.com/@moshe_31114/building-our-recycle-list-solution-in-react-17a21a9605a0
-   * @see https://dev.to/adamklein/build-your-own-virtual-scroll-part-i-11ib
-   * @see https://dev.to/adamklein/build-your-own-virtual-scroll-part-ii-3j86
-   * @see https://fluffy.es/solve-duplicated-cells
-   * @see https://vaadin.com/docs/latest/components/grid#columns
-   * @see https://www.htmlelements.com/demos/grid/datagrid-bind-to-json
-   * @see https://dev.to/gopal1996/understanding-reflow-and-repaint-in-the-browser-1jbg
-   * @see https://medium.com/teads-engineering/the-most-accurate-way-to-schedule-a-function-in-a-web-browser-eadcd164da12
-   * @see https://javascript.info/bubbling-and-capturing#capturing
-   * @see https://github.com/WICG/EventListenerOptions/blob/gh-pages/explainer.md
+   * Scroll a given row into view
+   * @param {number} rowIndex which row to scroll into view.
    */
-  scrollRowIntoView(rowIndex: number, doScroll = true) {
+  scrollRowIntoView(rowIndex: number) {
+    this.#scrollRowIntoView(rowIndex);
+  }
+
+  /**
+   * Scroll a given row into view
+   * @param {number} rowIndex which row to scroll into view.
+   * @param {boolean} doScroll set to "true" to have the browser perform the scroll action
+   */
+  #scrollRowIntoView(rowIndex: number, doScroll = true) {
     if (this.#rafReference) cancelAnimationFrame(this.#rafReference);
 
     const data = this.data;
@@ -2313,7 +2298,7 @@ export default class IdsDataGrid extends Base {
 
     let rowNode = this.rowByIndex(rowIndex);
     if (!rowNode && this.virtualScroll) {
-      this.scrollRowIntoView(rowIndex);
+      this.#scrollRowIntoView(rowIndex);
       rowNode = this.rowByIndex(rowIndex);
     }